Added header files for each class
added header files for Item class, Node class and LinkedList class and included them in main.cpp or customDB.cpp
