+++ CustomeDB/Item.h
@@ -1,0 +1,13 @@
+#pragma once
+#include <string>
+
+using namespace std;
+
+
+class Item {
+public:
+    string name;
+    int price;
+    int id;
+};
+
